Module oxilibc::stddef

stddef.h

Structs

max_align_t

A type whose alignment requirement is at least as strict (as large) as that of every scalar type.

Type Definitions

wchar_t

The integer type whose range of values can represent distinct codes for all members of the largest extended character set specified among the supported locales. The null character shall have the code value zero. Each member of the basic character set shall have a code value equal to its value when used as the lone character in an integer character constant if an implementation does not define STDC_MB_MIGHT_NEQ_WC.

wint_t

The integer type able to represent any value of the type wchar_t that is a member of the extended character set, as well as an additional value not part of that set.