Module oxilibc::stdint

stdint.h

Constants

INT8_MAX

INT8_MIN

INT8_WIDTH

INT16_MAX

INT16_MIN

INT16_WIDTH

INT32_MAX

INT32_MIN

INT32_WIDTH

INT64_MAX

INT64_MIN

INT64_WIDTH

INTMAX_MAX

INTMAX_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_FAST64_MAX

INT_FAST64_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

INT_LEAST64_MAX

INT_LEAST64_MIN

UINT8_MAX

UINT8_MIN

UINT8_WIDTH

UINT16_MAX

UINT16_MIN

UINT16_WIDTH

UINT32_MAX

UINT32_MIN

UINT32_WIDTH

UINT64_MAX

UINT64_MIN

UINT64_WIDTH

UINTMAX_MAX

UINTMAX_MIN

UINTPTR_MIN

UINT_FAST8_MAX

UINT_FAST8_MIN

UINT_FAST16_MAX

UINT_FAST16_MIN

UINT_FAST32_MAX

UINT_FAST32_MIN

UINT_FAST64_MAX

UINT_FAST64_MIN

UINT_LEAST8_MAX

UINT_LEAST8_MIN

UINT_LEAST16_MAX

UINT_LEAST16_MIN

UINT_LEAST32_MAX

UINT_LEAST32_MIN

UINT_LEAST64_MAX

UINT_LEAST64_MIN

Type Definitions

int_fast8_t

Fastest signed integer type with size of at least 8 bits

int_fast16_t

Fastest unsigned integer type with size of at least 16 bits

int_fast32_t

Fastest unsigned integer type with size of at least 32 bits

int_fast64_t

Fastest unsigned integer type with size of at least 64 bits

int_least8_t

Smallest signed integer type with size of 8 bits

int_least16_t

Smallest signed integer type with size of 16 bits

int_least32_t

Smallest signed integer type with size of 32 bits

int_least64_t

Smallest signed integer type with size of 64 bits

intmax_t

Maximum-size signed integer type on the platform

uint_fast8_t

Fastest unsigned integer type with size of at least 8 bits

uint_fast16_t

Fastest unsigned integer type with size of at least 16 bits

uint_fast32_t

Fastest unsigned integer type with size of at least 32 bits

uint_fast64_t

Fastest unsigned integer type with size of at least 64 bits

uint_least8_t

Smallest unsigned integer type with size of at least 8 bits

uint_least16_t

Smallest unsigned integer type with size of at least 16 bits

uint_least32_t

Smallest unsigned integer type with size of at least 32 bits

uint_least64_t

Smallest unsigned integer type with size of at least 64 bits

uintmax_t

Maximum-size unsigned integer type on the platform