Module oxilibc::string

string.h

Compliance

STD C < 11

memchr

memcmp

memset

memcpy

memmove

strcpy

strncpy

strcat

strncat

strxfrm

strlen

strcmp

strncmp

strcoll

strchr

strrchr

strspn

strcspn

strpbrk

strstr

strtok

strerror

STD C 11

memset_s

memcpy_s

memmove_s

strcpy_s

strncpy_s

strcat_s

strncat_s

strnlen_s

strtok_s

strerror_s

strerrorlen_s

STD C 23

memccpy

strdup

strndup

IEEE POSIX Std 1003.1-2017 (That are not part of current defined C stantards)

strnlen

stpcpy

stpncpy

strsignal

strcoll_l

strxfrm_l

strerror_l

strtok_r

strerror_r

Extensions

memrchr

memmem

memrmem

Maybe add extensions of memchr/memmem function family that returns the index of the found pointer

Functions

memccpy⚠

Copy bytes from src to dest until c byte is found, returning a pointer to the byte after the position of c in dest, or until len bytes are copied, returning a NULL pointer.

memchr⚠

Search for the first occurrence of a byte in a source first len bytes, returning a pointer to the found the byte or a null pointer it the byte was not found.

memcmp⚠

Compares the first len bytes of the objects pointed to by source and other and returning:

memcpy⚠

Copy the first len bytes from src to dest, returning dest itself.

memmem⚠

Search for the first occurrence of a pattern of pattern_len size in a src first len bytes, returning a pointer to the first byte of the found pattern or a null pointer it the pattern was not found.

memmove⚠

Copy the first len bytes from src to dest, returning dest.

memrchr⚠

Search for the last occurrence of a byte in a source first len bytes, returning a pointer to the found the byte or a null pointer it the byte was not found.

memrmem⚠

Search for the last occurrence of a pattern of pattern_len size in a src first len bytes, returning a pointer to the first byte of the found pattern or a null pointer it the pattern was not found.

memset⚠

Set the first len bytes of src to the value of val, returning the src.

stpcpy⚠

Copies the characters from the src string to the dest, returning a pointer to the end of the string, in other words, the first null character.

stpncpy⚠

Copies the at least max_size characters from the src string to the dest, returning a pointer to the end of the string, in other words, the first null character.

strcat⚠

Appends a copy of the src string to the dest string, returning a pointer to the beginning of the dest string.

strchr⚠

Search for the first occurrence of a ch in a str, returning a pointer to the found the ch or a null pointer it the ch was not found.

strcmp⚠

Compares str1 and str2 null terminated strings lexicographically. Returning:

strcpy⚠

Copies the src string, including the null terminator, to the dest, returning dest itself.

strcspn⚠

Returns the length of the maximum initial segment (span) of the byte string pointed to by str, that consists of only the characters not found in the byte string pointed to by char_set.

strlen⚠

Returns the length of the given null terminated string str.

strncat⚠

Appends a copy of the src string to the dest string for at least max_size characters, returning a pointer to the beginning of the dest string.

strncmp⚠

Compares str1 and str2 for a maximum of max_size strings lexicographically. Returning:

strncpy⚠

Copies the at least max_size characters from the src string to the dest, returning dest itself.

strnlen⚠

A safer version of strlen. This function returns:

strnlen_s⚠

A safer version of strlen. This function returns:

strpbrk⚠

Search for the first ocorrence of any character of break_set in the str, returning a pointer to the first character ocurrence or a null pointer if none found.

strrchr⚠

Search for the last occurrence of a ch in a str, returning a pointer to the found the ch or a null pointer it the ch was not found.

strspn⚠

Returns the length of the maximum initial segment (span) of the byte string pointed to by str, that consists of only the characters found in the byte string pointed to by char_set.

strstr⚠

Search for the first occurrence of a pattern in a src, returning a pointer to the first character of the found the pattern or a null pointer it the pattern was not found.