Function oxilibc::string::memccpy

#[no_mangle]
pub unsafe extern "C" fn memccpy(
    dest: *mut c_void,
    src: *const c_void,
    c: c_int,
    len: size_t
) -> *mut c_void

Copy bytes from src to dest until c byte is found, returning a pointer to the byte after the position of c in dest, or until len bytes are copied, returning a NULL pointer.

Safety

Behavior is undefined if any of the following conditions are violated:

src must be valid for reads of len bytes.

dest must be valid for writes of len bytes.

The region of memory beginning at src with a size of len bytes must not overlap with the region of memory beginning at dest with the same size.

c must be a valid u8 value.

Examples

Rust

use oxilibc::string::memccpy;

let src = [1u8, 2, 3, 4, 5];
let mut dest = [0u8; 5];

let ptr = unsafe { memccpy(dest.as_mut_ptr() as _, src.as_ptr() as _, 3, 5) };
assert_eq!(ptr, dest[3..].as_ptr() as _);
assert_eq!(dest, [1u8, 2, 3, 0, 0]);

use oxilibc::string::memccpy;

let src = [1u8, 2, 3, 4, 5];
let mut dest = [0u8; 5];

let ptr = unsafe { memccpy(dest.as_mut_ptr() as _, src.as_ptr() as _, 7, 5) };
assert!(ptr.is_null());
assert_eq!(dest, [1u8, 2, 3, 4, 5]);

C

#include <string.h>
#include <assert.h>

int main(void) {
    const char src[] = {1, 2, 3, 4, 5};
    char* dest = {0,0,0,0,0};

    void* ptr = memccpy(dest, src, 3, 5);
    assert(memcmp(dest, src, 3));

    return 0;
}

#include <string.h>
#include <assert.h>

int main(void) {
    const char src[] = {1, 2, 3, 4, 5};
    char* dest = {0,0,0,0,0};

    void* ptr = memccpy(dest, src, 7, 5);
    assert(NULL == ptr);
    assert(memcmp(dest, src, 5));

    return 0;
}