Function oxilibc::string::memcmp

#[no_mangle]
pub unsafe extern "C" fn memcmp(
    src: *const c_void,
    other: *const c_void,
    len: size_t
) -> c_int

Compares the first len bytes of the objects pointed to by source and other and returning:

Zero if both pointed objects are the same

A positive number if source is greater than other

A negative number if source is less than other

Safety

Behavior is undefined if any of the following conditions are violated:

src must be valid for reads of len bytes.

other must be valid for reads of len bytes.

Examples

Rust

use oxilibc::string::memcmp;

let source = b"Hello,";
let other = b"World";

let len = source.len().min(other.len());
let ord = unsafe { memcmp(source.as_ptr() as _, other.as_ptr() as _, len) };

assert!(ord < 0);

C

#include <string.h>
#include <assert.h>

int main(void) {
    const char source[] = "Hello,";
    const char other[]  = "World";

    int ord = memcmp(source, other, 5);
    assert(ord < 0);

    return 0;
}