Function oxilibc::string::memmove

#[no_mangle]
pub unsafe extern "C" fn memmove(
    dest: *mut c_void,
    src: *const c_void,
    len: size_t
) -> *mut c_void

Copy the first len bytes from src to dest, returning dest.

Note that differently than memcpy, the slices of memory of [src, src + len) and [dest, dest + len) can overlap.

Safety

Behavior is undefined if any of the following conditions are violated:

src must be valid for reads of len bytes.

dest must be valid for writes of len bytes.

Examples

Rust

use oxilibc::string::memmove;

let mut src = [1u8, 2, 3, 4, 5];
let src2 = [125u8, 124, 122];
let p1 = src.as_mut_ptr();
let p2 = src2.as_ptr();
unsafe { memmove(p1.cast(), p2.cast(), src2.len() as _) };
assert_eq!(src, [125, 124, 122, 4, 5]);

C

#include <string.h>
#include <assert.h>

int main(void) {
    char* src = "abcde";
    char* dest = "FGH";

    assert(memcmp(memmove(dest, src, 3), "FGHde", 6));

    return 0;
}