Function oxilibc::string::memrchr

#[no_mangle]
pub unsafe extern "C" fn memrchr(
    src: *const c_void,
    byte: c_int,
    len: size_t
) -> *mut c_void

Search for the last occurrence of a byte in a source first len bytes, returning a pointer to the found the byte or a null pointer it the byte was not found.

Safety

Behavior is undefined if any of the following conditions are violated:

src must be valid for reads of len bytes.

Examples

Rust

use oxilibc::string::memrchr;

let arr = [10, 12, 45, 12, 32];

let found_at_ptr = unsafe { memrchr(arr.as_ptr() as _, 12, arr.len()) };

assert_eq!(found_at_ptr, arr[3..].as_ptr() as _);

C

#include <string.h>
#include <assert.h>

int main(void) {
    const char arr[] = {10, 12, 45, 12, 32};

    char* ptr = memrchr(arr, 12, 5);
    assert(ptr == arr[3]);

    return 0;
}