Function oxilibc::string::stpncpy

#[no_mangle]
pub unsafe extern "C" fn stpncpy(
    dest: *mut c_char,
    src: *const c_char,
    max_size: size_t
) -> *mut c_char

Copies the at least max_size characters from the src string to the dest, returning a pointer to the end of the string, in other words, the first null character.

Safety

Behavior is undefined if any of the following conditions are violated:

src must be a null terminated string

dest must be at least the size of max_size bytes long or the size of src if it is smaller

str must be valid for reads for max_size bytes or until the first null character found, whatever is smaller.

dest must be valid for writes for max_size bytes or until the str length, whatever is smaller.

src and dest must not overlap

Examples

Rust

use oxibot::string::stpcpy;

let src = b"Hello World!\0";
let dest = &mut [0u8; 13];
let ret = unsafe { stpncpy(dest.as_mut_ptr().cast(), src.as_ptr().cast(), 5) };
assert_eq!(b"Hello\0\0\0\0\0\0\0\0", dest);

let expected_ret = unsafe { dest.as_mut_ptr().add(5).cast() };
assert_eq!(ret, expected_ret);

unsafe {
    assert_eq!(*ret, 0);
};

C

#include <string.h>
#include <assert.h>

int main(void) {
    const char* src = "Hello world!";
    char dest[13] = {0};

    char* last = stpncpy(dest, src, 5);

    for (int i = 0; i < 5; i++) {
        assert(src[i] == dest[i]);
    }

    assert(*last == 0);

    return 0;
}