Function oxilibc::string::strpbrk

#[no_mangle]
pub unsafe fn strpbrk(
    str: *const c_char,
    break_set: *const c_char
) -> *mut c_char

Search for the first ocorrence of any character of break_set in the str, returning a pointer to the first character ocurrence or a null pointer if none found.

Safety

Behavior is undefined if any of the following conditions are violated:

str must be a null terminated string

break_set must be a null terminated string

str must be valid for reads until the first null character found

break_set must be valid for reads until the first null character found.

Examples

use oxibot::string::strpbrk;

let s = b"abcde312$#@\0";
let pat = b"3\0";

let test = unsafe { strpbrk(s.as_ptr().cast(), pat.as_ptr().cast()) };
let ans = s[5..].as_ptr().cast::<c_char>() as *mut _;
assert_eq!(test, ans);

#include <string.h>
#include <assert.h>

int main(void) {
    const char* string = "abcde312$#@";
    const char* break_set = "3";

    char* slice = strpbrk(string, break_set);
    assert(slice == string[5]);

    return 0;
}