Function oxilibc::string::strrchr

#[no_mangle]
pub unsafe extern "C" fn strrchr(
    str: *const c_char,
    ch: c_int
) -> *mut c_char

Search for the last occurrence of a ch in a str, returning a pointer to the found the ch or a null pointer it the ch was not found.

Safety

Behavior is undefined if any of the following conditions are violated:

str must be a null terminated string

str must be valid for reads until the first null character found.

Examples

Rust

use oxilibc::string::strrchr;

let str = b"Hello, World!\0";

let found_at_ptr = unsafe { strrchr(str.as_ptr() as _, b'l' as _) };

assert_eq!(found_at_ptr, str[10..].as_ptr() as _);

C

#include <string.h>
#include <assert.h>

int main(void) {
    const char str[] = "Hello, World!";

    char* ptr = memchr(str, 'l');
    assert(ptr == str[10]);

    return 0;
}