Function oxilibc::string::strstr

#[no_mangle]
pub unsafe extern "C" fn strstr(
    str: *const c_char,
    pattern: *const c_char
) -> *mut c_char

Search for the first occurrence of a pattern in a src, returning a pointer to the first character of the found the pattern or a null pointer it the pattern was not found.

Safety

Behavior is undefined if any of the following conditions are violated:

src and pattern must be null terminated strings

src must be valid for reads until the first null character found.

pattern must be valid for reads until the first null character found.

Examples

Rust

use oxibot::string::strstr;

let src = b"Hello World!\0";
let pat = b"World\0";

let res = unsafe { strstr(src.as_ptr().cast(), pat.as_ptr().cast()) };
assert_eq!(unsafe { src.as_ptr().cast::<i8>().add(6) }, res);

C

#include <string.h>
#include <assert.h>

int main(void) {
    const char* src = "Hello world!";
    const char* pat = "world";

    char* res = strstr(src, pat);

    assert(src[6] == res);

    return 0;
}