Module oxilibc::strings

strings.h

Although not part of the Standard C, it is part of POSIX Standard, but most functions are generic enough to make then available on all platforms

Compliance

IEEE POSIX Std 1003.1-2008

bcmp

bcopy

bzero

ffs

strcasecmp

strncasecmp

index

rindex

IEEE POSIX Std 1003.1-2017

strcasecmp_l

strncasecmp_l

Extensions

ffsl

ffsll

fls

flsl

flsll

Functions

bcmp⚠Deprecated

This function is the same as memcmp function defined on the string.h. For all intents and purposes, use memcmp instead.

bcopy⚠Deprecated

This function is the same as memmove function defined on the string.h. For all intents and purposes, use memmove instead.

bzero⚠Deprecated

This function is the same as memset function defined on the string.h with val equal to zero. For all intents and purposes, use memset instead.

ffs

Finds the first bit set (beginning with the least significant bit) and returns the index of that bit of the given word.

ffsl

Finds the first bit set (beginning with the least significant bit) and returns the index of that bit of the given word.

ffsll

Finds the first bit set (beginning with the least significant bit) and returns the index of that bit of the given word.

index⚠Deprecated

This function is the same as strchr function defined on the string.h with val equal to zero. For all intents and purposes, use strchr instead.

rindex⚠Deprecated

This function is the same as strrchr function defined on the string.h with val equal to zero. For all intents and purposes, use strrchr instead.

strcasecmp⚠

Compares the string at str1 to the string at str2 in a case-insensitive manner, returning:

strncasecmp⚠

Compares the string at str1 to the string at str2 in a case-insensitive manner until the len or one of the strings get to the end, returning: