Function oxilibc::strings::bcmp

#[no_mangle]
pub unsafe extern "C" fn bcmp(
    source: *const c_void,
    other: *const c_void,
    len: size_t
) -> c_int

👎Deprecated: This function exist only for compatibility with older POSIX systems. Use memcmp instead

This function is the same as memcmp function defined on the string.h. For all intents and purposes, use memcmp instead.

Safety

The user of this function must hold these invariants for the function to not cause Undefined Behavior:

source and other must be a valid pointer

len value must be between zero and the the minimum size between source and other