Function oxilibc::strings::strncasecmp

#[no_mangle]
pub unsafe extern "C" fn strncasecmp(
    src: *const c_char,
    other: *const c_char,
    len: size_t
) -> c_int

Compares the string at str1 to the string at str2 in a case-insensitive manner until the len or one of the strings get to the end, returning:

Zero if both strings match

A positive number if str1 sorts lexicographically after str2

A negative number if str1 sorts lexicographically before str2

Safety

Behavior is undefined if any of the following conditions are violated:

src must be valid for reads of len bytes.

src must be null terminated and valid for reads until the null temination byte.

other must be valid for reads of len bytes.

other must be null terminated and valid for reads until the null temination byte.

Examples

Rust:

use oxilibc::strings::strncasecmp;

let (s1, s2, s3, s4) = (b"Foofoo\0", b"fooFoo\0", b"foobar\0", b"barFoo\0");

unsafe {
    // the same case insensitive
    assert_eq!(strncasecmp(s1.as_ptr().cast(), s2.as_ptr().cast(), 6), 0);
    assert_eq!(strncasecmp(s1.as_ptr().cast(), s3.as_ptr().cast(), 3), 0);

    assert!(strncasecmp(s1.as_ptr().cast(), s4.as_ptr().cast(), 6) > 0);
    assert!(strncasecmp(s4.as_ptr().cast(), s1.as_ptr().cast(), 6) < 0);
}

C:

#include <strings.h>
#include <assert.h>

int main(void) {
    const char* s1 = "Foofoo";
    const char* s2 = "fooFoo";
    const char* s3 = "foobar";
    const char* s4 = "barFoo";

    assert(strncasecmp(s1, s2, 6) == 0);
    assert(strncasecmp(s1, s3, 3) == 0);
    assert(strncasecmp(s1, s4, 6) > 0);
    assert(strncasecmp(s4, s1, 6) < 0);

    return 0;
}